Final refactoring and cleanup
