+++ src/main/java/org/polypheny/simpleclient/scenario/graph/queryBuilder/CreateGraphDatabase.java
@@ -27,14 +27,14 @@ package org.polypheny.simpleclient.scenario.graph.queryBuilder;
 import org.polypheny.simpleclient.query.Query;
 import org.polypheny.simpleclient.query.QueryBuilder;
 import org.polypheny.simpleclient.scenario.graph.GraphBench;
-import org.polypheny.simpleclient.scenario.graph.queryBuilder.GraphInsert.SimpleGraphInsert;
+import org.polypheny.simpleclient.scenario.graph.GraphInsert;
 
 public class CreateGraphDatabase extends QueryBuilder {
 
 
     @Override
     public Query getNewQuery() {
-        return new SimpleGraphInsert( String.format( "CREATE DATABASE %s IF NOT EXISTS", GraphBench.GRAPH_DATABASE ) );
+        return new GraphInsert( String.format( "CREATE DATABASE %s IF NOT EXISTS", GraphBench.GRAPH_DATABASE ) );
     }
 
 }
