+++ src/main/java/org/polypheny/simpleclient/scenario/graph/queryBuilder/DifferentPathsBuilder.java
@@ -28,7 +28,7 @@ import java.util.Random;
 import org.polypheny.simpleclient.query.Query;
 import org.polypheny.simpleclient.query.QueryBuilder;
 import org.polypheny.simpleclient.scenario.graph.GraphBenchConfig;
-import org.polypheny.simpleclient.scenario.graph.GraphQuery.SimpleGraphQuery;
+import org.polypheny.simpleclient.scenario.graph.GraphQuery;
 
 public class DifferentPathsBuilder extends QueryBuilder {
 
@@ -44,7 +44,7 @@ public class DifferentPathsBuilder extends QueryBuilder {
 
     @Override
     public Query getNewQuery() {
-        return new SimpleGraphQuery( String.format( "MATCH (n:Label0)-[*%s]-(p) RETURN p", random.nextInt( config.maxPathLength - 2 ) + 1 ) );
+        return new GraphQuery( String.format( "MATCH (n:Label0)-[*%s]-(p) RETURN p", random.nextInt( config.maxPathLength - 2 ) + 1 ) );
     }
 
 }
